+++ server/src/test/scala/it/unibo/scalapacman/server/config/TestSettings.scala
@@ -1,4 +1,4 @@
-package it.unibo.scalapacman.server.util
+package it.unibo.scalapacman.server.config
 
 import scala.concurrent.duration.FiniteDuration
