+++ server/src/test/scala/it/unibo/scalapacman/server/core/MasterTest.scala
@@ -5,7 +5,7 @@ import java.util.concurrent.TimeUnit
 import akka.actor.testkit.typed.scaladsl.{ScalaTestWithActorTestKit, TestProbe}
 import akka.actor.typed.ActorRef
 import akka.actor.typed.receptionist.{Receptionist, ServiceKey}
-import it.unibo.scalapacman.server.util.ConfLoader
+import it.unibo.scalapacman.server.config.ConfLoader
 import org.scalatest.wordspec.AnyWordSpecLike
 
 import scala.concurrent.duration.FiniteDuration
